export error fixed + telegram plot

+++ app/routes/analyze.py
@@ -48,25 +48,25 @@ def analyze_data():
         "start": start.isoformat(),
         "end": end.isoformat(),
 
-        "avg_temp": round(df['temperature'].mean(), 2),
-        "max_temp": round(df['temperature'].max(), 2),
-        "min_temp": round(df['temperature'].min(), 2),
-        "median_temp": round(df['temperature'].median(), 2),
-        "temp_range": round(df['temperature'].max() - df['temperature'].min(), 2),
-        "temp_exceeds_count": len(temp_exceeds),
+        "avg_temp": round(float(df['temperature'].mean()), 2),
+        "max_temp": round(float(df['temperature'].max()), 2),
+        "min_temp": round(float(df['temperature'].min()), 2),
+        "median_temp": round(float(df['temperature'].median()), 2),
+        "temp_range": round(float(df['temperature'].max() - df['temperature'].min()), 2),
+        "temp_exceeds_count": int(len(temp_exceeds)),
 
-        "avg_humidity": round(df['humidity'].mean(), 2),
-        "max_humidity": round(df['humidity'].max(), 2),
-        "min_humidity": round(df['humidity'].min(), 2),
-        "median_humidity": round(df['humidity'].median(), 2),
-        "humidity_range": round(df['humidity'].max() - df['humidity'].min(), 2),
-        "humidity_exceeds_count": len(humidity_exceeds),
+        "avg_humidity": round(float(df['humidity'].mean()), 2),
+        "max_humidity": round(float(df['humidity'].max()), 2),
+        "min_humidity": round(float(df['humidity'].min()), 2),
+        "median_humidity": round(float(df['humidity'].median()), 2),
+        "humidity_range": round(float(df['humidity'].max() - df['humidity'].min()), 2),
+        "humidity_exceeds_count": int(len(humidity_exceeds)),
 
-        "std_light": round(df['light'].std(), 2),
-        "max_light": round(df['light'].max(), 2),
-        "min_light": round(df['light'].min(), 2),
-        "light_range": round(df['light'].max() - df['light'].min(), 2),
-        "light_exceeds_count": len(light_exceeds)
+        "std_light": round(float(df['light'].std()), 2),
+        "max_light": round(float(df['light'].max()), 2),
+        "min_light": round(float(df['light'].min()), 2),
+        "light_range": round(float(df['light'].max() - df['light'].min()), 2),
+        "light_exceeds_count": int(len(light_exceeds))
     }
 
     return jsonify(analysis)
